refactor coarsening: extract concurrency logic (#228)
