+++ mt-kahypar/partition/coarsening/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(MultilevelCoarseningSources
         multilevel/deterministic_multilevel_coarsener.cpp
-        multilevel/multilevel_uncoarsener.cpp)
+        multilevel/multilevel_uncoarsener.cpp
+        multilevel/concurrent_clustering_data.cpp)
 
 set(NLevelCoarseningSources
         nlevel/nlevel_uncoarsener.cpp)
