+++ mt-kahypar/partition/coarsening/multilevel/multilevel_coarsener.h
@@ -29,8 +29,6 @@
 
 #include <string>
 
-#include <tbb/concurrent_queue.h>
-#include <tbb/task_group.h>
 #include <tbb/parallel_for.h>
 #include <tbb/parallel_reduce.h>
 
@@ -38,13 +36,13 @@
 
 #include "include/mtkahypartypes.h"
 
+#include "mt-kahypar/partition/coarsening/multilevel/concurrent_clustering_data.h"
 #include "mt-kahypar/partition/coarsening/multilevel/multilevel_coarsener_base.h"
 #include "mt-kahypar/partition/coarsening/multilevel/multilevel_vertex_pair_rater.h"
 #include "mt-kahypar/partition/coarsening/i_coarsener.h"
 #include "mt-kahypar/partition/coarsening/policies/rating_acceptance_policy.h"
 #include "mt-kahypar/partition/coarsening/policies/rating_heavy_node_penalty_policy.h"
 #include "mt-kahypar/partition/coarsening/policies/rating_score_policy.h"
-#include "mt-kahypar/parallel/atomic_wrapper.h"
 #include "mt-kahypar/utils/cast.h"
 #include "mt-kahypar/utils/progress_bar.h"
 #include "mt-kahypar/utils/randomize.h"
@@ -61,23 +59,9 @@ class MultilevelCoarsener : public ICoarsener,
  private:
 
   using Base = MultilevelCoarsenerBase<TypeTraits>;
-  using Rater = MultilevelVertexPairRater<ScorePolicy,
-                                          HeavyNodePenaltyPolicy,
-                                          AcceptancePolicy>;
-  using Rating = typename Rater::Rating;
+  using Rating = MultilevelVertexPairRater::Rating;
   using Hypergraph = typename TypeTraits::Hypergraph;
   using PartitionedHypergraph = typename TypeTraits::PartitionedHypergraph;
-
-  enum class MatchingState : uint8_t {
-    UNMATCHED = 0,
-    MATCHING_IN_PROGRESS = 1,
-    MATCHED = 2
-  };
-
-  #define STATE(X) static_cast<uint8_t>(X)
-  using AtomicMatchingState = parallel::IntegralAtomicWrapper<uint8_t>;
-  using AtomicWeight = parallel::IntegralAtomicWrapper<HypernodeWeight>;
-  using AtomicID = parallel::IntegralAtomicWrapper<HypernodeID>;
 
   static constexpr bool debug = false;
   static constexpr bool enable_heavy_assert = false;
@@ -94,36 +78,18 @@ class MultilevelCoarsener : public ICoarsener,
            utils::cast<Hypergraph>(hypergraph).maxEdgeSize(), context),
     _initial_num_nodes(utils::cast<Hypergraph>(hypergraph).initialNumNodes()),
     _current_vertices(),
-    _matching_state(),
-    _cluster_weight(),
-    _matching_partner(),
+    _clustering_data(_hg.initialNumNodes(), context),
     _pass_nr(0),
     _progress_bar(utils::cast<Hypergraph>(hypergraph).initialNumNodes(), 0, false),
     _enable_randomization(true) {
     _progress_bar += _hg.numRemovedHypernodes();
-
-    // Initialize internal data structures parallel
-    tbb::parallel_invoke([&] {
-      _current_vertices.resize(_hg.initialNumNodes());
-    }, [&] {
-      _matching_state.resize(_hg.initialNumNodes());
-    }, [&] {
-      _cluster_weight.resize(_hg.initialNumNodes());
-    }, [&] {
-      _matching_partner.resize(_hg.initialNumNodes());
-    });
+    _current_vertices.resize(_hg.initialNumNodes());
   }
 
   MultilevelCoarsener(const MultilevelCoarsener&) = delete;
   MultilevelCoarsener(MultilevelCoarsener&&) = delete;
   MultilevelCoarsener & operator= (const MultilevelCoarsener &) = delete;
   MultilevelCoarsener & operator= (MultilevelCoarsener &&) = delete;
-
-  ~MultilevelCoarsener() {
-    parallel::parallel_free(
-      _current_vertices, _matching_state,
-      _cluster_weight, _matching_partner);
-  }
 
   void disableRandomization() {
     _enable_randomization = false;
@@ -150,18 +116,11 @@ class MultilevelCoarsener : public ICoarsener,
 
     // Random shuffle vertices of current hypergraph
     _current_vertices.resize(current_hg.initialNumNodes());
-    parallel::scalable_vector<HypernodeID> cluster_ids(current_hg.initialNumNodes());
+    parallel::scalable_vector<HypernodeID> cluster_ids;
     tbb::parallel_for(ID(0), current_hg.initialNumNodes(), [&](const HypernodeID hn) {
-      ASSERT(hn < _current_vertices.size());
-      // Reset clustering
       _current_vertices[hn] = hn;
-      _matching_state[hn].store(STATE(MatchingState::UNMATCHED), std::memory_order_relaxed);
-      _matching_partner[hn].store(hn, std::memory_order_relaxed);
-      cluster_ids[hn] = hn;
-      if ( current_hg.nodeIsEnabled(hn) ) {
-        _cluster_weight[hn] = current_hg.nodeWeight(hn);
-      }
     });
+    _clustering_data.initializeCoarseningPass(current_hg, cluster_ids);
 
     if ( _enable_randomization ) {
       utils::Randomize::instance().parallelShuffleVector( _current_vertices, UL(0), _current_vertices.size());
@@ -177,32 +136,8 @@ class MultilevelCoarsener : public ICoarsener,
     }
     DBG << V(current_num_nodes);
 
-    HEAVY_COARSENING_ASSERT([&] {
-      parallel::scalable_vector<HypernodeWeight> expected_weights(current_hg.initialNumNodes());
-      // Verify that clustering is correct
-      for ( const HypernodeID& hn : current_hg.nodes() ) {
-        const HypernodeID u = hn;
-        const HypernodeID root_u = cluster_ids[u];
-        if ( root_u != cluster_ids[root_u] ) {
-          LOG << "Hypernode" << u << "is part of cluster" << root_u << ", but cluster"
-              << root_u << "is also part of cluster" << cluster_ids[root_u];
-          return false;
-        }
-        expected_weights[root_u] += current_hg.nodeWeight(hn);
-      }
-
-      // Verify that cluster weights are aggregated correct
-      for ( const HypernodeID& hn : current_hg.nodes() ) {
-        const HypernodeID u = hn;
-        const HypernodeID root_u = cluster_ids[u];
-        if ( root_u == u && expected_weights[u] != _cluster_weight[u] ) {
-          LOG << "The expected weight of cluster" << u << "is" << expected_weights[u]
-              << ", but currently it is" << _cluster_weight[u];
-          return false;
-        }
-      }
-      return true;
-    }(), "Parallel clustering computed invalid cluster ids and weights");
+    HEAVY_COARSENING_ASSERT(_clustering_data.verifyClustering(current_hg, cluster_ids),
+                            "Parallel clustering computed invalid cluster ids and weights");
 
     const double reduction_vertices_percentage =
       static_cast<double>(num_hns_before_pass) /
@@ -250,39 +185,40 @@ class MultilevelCoarsener : public ICoarsener,
         //  1.) The contraction limit of the current level is not reached
         //  2.) Vertex hn is not matched before
         const HypernodeID u = hn;
-        if (_matching_state[u] == STATE(MatchingState::UNMATCHED)) {
-          if (current_num_nodes > hierarchy_contraction_limit) {
-            ASSERT(current_hg.nodeIsEnabled(hn));
-            const Rating rating = _rater.template rate<has_fixed_vertices>(current_hg, hn,
-              cluster_ids, _cluster_weight, fixed_vertices, _context.coarsening.max_allowed_node_weight);
-            if (rating.target != kInvalidHypernode) {
-              const HypernodeID v = rating.target;
-              HypernodeID& local_contracted_nodes = contracted_nodes.local();
-              matchVertices<has_fixed_vertices>(current_hg, u, v,
-                cluster_ids, local_contracted_nodes, fixed_vertices);
+        if (_clustering_data.vertexIsUnmatched(u) && current_num_nodes > hierarchy_contraction_limit) {
+          ASSERT(current_hg.nodeIsEnabled(hn));
+          const Rating rating = _rater.template rate<ScorePolicy, HeavyNodePenaltyPolicy, AcceptancePolicy, has_fixed_vertices>(
+            current_hg, hn, cluster_ids, _clustering_data.clusterWeight(), fixed_vertices, _context.coarsening.max_allowed_node_weight);
+          if (rating.target != kInvalidHypernode) {
+            const HypernodeID v = rating.target;
+            HypernodeID& local_contracted_nodes = contracted_nodes.local();
+            bool success = _clustering_data.template matchVertices<has_fixed_vertices>(
+              current_hg, u, v, cluster_ids, _rater, fixed_vertices);
+            if (success) {
+              ++local_contracted_nodes;
+            }
 
-              // To maintain the current number of nodes of the hypergraph each PE sums up
-              // its number of contracted nodes locally. To compute the current number of
-              // nodes, we have to sum up the number of contracted nodes of each PE. This
-              // operation becomes more expensive the more PEs are participating in coarsening.
-              // In order to prevent expensive updates of the current number of nodes, we
-              // define a threshold which the local number of contracted nodes have to exceed
-              // before the current PE updates the current number of nodes. This threshold is defined
-              // by the distance to the current contraction limit divided by the number of PEs.
-              // Once one PE exceeds this bound the first time it is not possible that the
-              // contraction limit is reached, because otherwise an other PE would update
-              // the global current number of nodes before. After update the threshold is
-              // increased by the new difference (in number of nodes) to the contraction limit
-              // divided by the number of PEs.
-              if (local_contracted_nodes >= num_nodes_update_threshold.local()) {
-                current_num_nodes = num_hns_before_pass -
-                                    contracted_nodes.combine(std::plus<HypernodeID>());
-                const HypernodeID dist_to_contraction_limit =
-                  current_num_nodes > hierarchy_contraction_limit ?
-                  current_num_nodes - hierarchy_contraction_limit : 0;
-                num_nodes_update_threshold.local() +=
-                  dist_to_contraction_limit / _context.shared_memory.original_num_threads;
-              }
+            // To maintain the current number of nodes of the hypergraph each PE sums up
+            // its number of contracted nodes locally. To compute the current number of
+            // nodes, we have to sum up the number of contracted nodes of each PE. This
+            // operation becomes more expensive the more PEs are participating in coarsening.
+            // In order to prevent expensive updates of the current number of nodes, we
+            // define a threshold which the local number of contracted nodes have to exceed
+            // before the current PE updates the current number of nodes. This threshold is defined
+            // by the distance to the current contraction limit divided by the number of PEs.
+            // Once one PE exceeds this bound the first time it is not possible that the
+            // contraction limit is reached, because otherwise an other PE would update
+            // the global current number of nodes before. After update the threshold is
+            // increased by the new difference (in number of nodes) to the contraction limit
+            // divided by the number of PEs.
+            if (local_contracted_nodes >= num_nodes_update_threshold.local()) {
+              current_num_nodes = num_hns_before_pass -
+                                  contracted_nodes.combine(std::plus<HypernodeID>());
+              const HypernodeID dist_to_contraction_limit =
+                current_num_nodes > hierarchy_contraction_limit ?
+                current_num_nodes - hierarchy_contraction_limit : 0;
+              num_nodes_update_threshold.local() +=
+                dist_to_contraction_limit / _context.shared_memory.original_num_threads;
             }
           }
         }
@@ -340,152 +276,6 @@ class MultilevelCoarsener : public ICoarsener,
     _uncoarseningData.finalizeCoarsening();
   }
 
-  /*!
-   * We maintain the invariant during clustering that each cluster has a unique
-   * representative and all vertices also part of that cluster point to that
-   * representative. Let v be the representative of a cluster C_v, then for
-   * all nodes u \in C_v follows that cluster_ids[u] = v.
-   * If we perform sequential clustering, we can simply set
-   * cluster_ids[u] = cluster_ids[v] to maintain our invariant. However,
-   * things become more complicated if we perform parallel clustering.
-   * Especially, if two neighbors u and v are concurrently matched, we have
-   * to guarantee that our clustering fullfils our invariant. There are mainly
-   * two different cases, which needs special attention:
-   *   1.) u is matched with v and v is matched with u concurrently
-   *   2.) u is matched with v and v is matched an other vertex w concurrently
-   * The following functions guarantees that our invariant is fullfilled, if
-   * vertices are matched concurrently.
-   */
-  template<bool has_fixed_vertices>
-  MT_KAHYPAR_ATTRIBUTE_ALWAYS_INLINE bool matchVertices(const Hypergraph& hypergraph,
-                                                        const HypernodeID u,
-                                                        const HypernodeID v,
-                                                        parallel::scalable_vector<HypernodeID>& cluster_ids,
-                                                        HypernodeID& contracted_nodes,
-                                                        ds::FixedVertexSupport<Hypergraph>& fixed_vertices) {
-    // MEMORY ORDERING AND SYNCHRONIZATION
-    // During the clustering, there are concurrent memory accesses to the following 4 locations:
-    // _matching_state, cluster_ids, _matching_partner and _cluster_weight.
-    // We use _matching_state to synchronize accesses to cluster_ids with acquire/release semantics, while
-    // _matching_partner and _cluster_weight don't require synchronization. In more detail (see also PR #193):
-    //   1. We read cluster_ids[v] to determine the representative for joinCluster if v is already matched.
-    //      This is synchronized by using release ordering when writing the MATCHED state to _matching_state
-    //      and always checking _matching_state[v] with acquire ordering before accessing cluster_ids[v]
-    //   2. _matching_partner is used for conflict resolution. Since the conflict resolution loops until
-    //      a stable state is detected, no explicit synchronization is necessary and relaxed ordering suffices
-    //   3. Updating _cluster_weight might cause a race condition in joinCluster. However, this just causes the
-    //      cluster to exceed the allowed weight. This is acceptable since it rarely happens in practice
-
-    ASSERT(u < hypergraph.initialNumNodes());
-    ASSERT(v < hypergraph.initialNumNodes());
-    const uint8_t matched = STATE(MatchingState::MATCHED);
-    const uint8_t match_in_progress = STATE(MatchingState::MATCHING_IN_PROGRESS);
-
-    // Indicates that u wants to join the cluster of v.
-    // Will be important later for conflict resolution.
-    bool success = false;
-    const HypernodeWeight weight_u = hypergraph.nodeWeight(u);
-    HypernodeWeight weight_v = _cluster_weight[v].load(std::memory_order_relaxed);
-    if ( weight_u + weight_v <= _context.coarsening.max_allowed_node_weight ) {
-      uint8_t expect_unmatched_u = STATE(MatchingState::UNMATCHED);
-      if ( _matching_state[u].compare_exchange_strong(expect_unmatched_u, match_in_progress, std::memory_order_relaxed) ) {
-        _matching_partner[u].store(v, std::memory_order_relaxed);
-        // Current thread gets "ownership" for vertex u. Only threads with "ownership"
-        // can change the cluster id of a vertex.
-
-        uint8_t expect_unmatched_v = STATE(MatchingState::UNMATCHED);
-        uint8_t matching_state_v = _matching_state[v].load(std::memory_order_acquire);
-        if ( matching_state_v == matched ) {
-          // Vertex v is already matched and will not change it cluster id any more.
-          // In that case, it is safe to set the cluster id of u to the cluster id of v.
-          const HypernodeID rep = cluster_ids[v];
-          ASSERT(_matching_state[rep] == matched);
-          success = joinCluster<has_fixed_vertices>(hypergraph,
-            u, rep, cluster_ids, contracted_nodes, fixed_vertices);
-        } else if ( matching_state_v == expect_unmatched_v &&
-                    _matching_state[v].compare_exchange_strong(expect_unmatched_v, match_in_progress, std::memory_order_relaxed) ) {
-          // Current thread has the "ownership" for u and v and can change the cluster id
-          // of both vertices thread-safe.
-          success = joinCluster<has_fixed_vertices>(hypergraph,
-            u, v, cluster_ids, contracted_nodes, fixed_vertices);
-          _matching_state[v].store(matched, std::memory_order_release);
-        } else {
-          // State of v must be either MATCHING_IN_PROGRESS or an other thread changed the state
-          // in the meantime to MATCHED. We have to wait until the state of v changed to
-          // MATCHED or resolve the conflict if u is matched within a cyclic matching dependency
-
-          // Conflict Resolution
-          do {
-            // Check if current vertex is in a cyclic matching dependency
-            HypernodeID cur_u = u;
-            HypernodeID smallest_node_id_in_cycle = cur_u;
-            while (true) {
-              HypernodeID next_u = _matching_partner[cur_u].load(std::memory_order_relaxed);
-              if (next_u != u && next_u != cur_u) {
-                cur_u = next_u;
-                smallest_node_id_in_cycle = std::min(smallest_node_id_in_cycle, cur_u);
-              } else {
-                break;
-              }
-            }
-
-            // Resolve cyclic matching dependency
-            // Vertex with smallest id starts to resolve conflict
-            const bool is_in_cyclic_dependency = _matching_partner[cur_u].load(std::memory_order_relaxed) == u;
-            if ( is_in_cyclic_dependency && u == smallest_node_id_in_cycle) {
-              success = joinCluster<has_fixed_vertices>(hypergraph,
-                u, v, cluster_ids, contracted_nodes, fixed_vertices);
-              _matching_state[v].store(matched, std::memory_order_release);
-            }
-          } while ( _matching_state[v].load(std::memory_order_acquire) == match_in_progress );
-          // note: the loop provides acquire semantics for the block below
-
-          // If u is still in state MATCHING_IN_PROGRESS its matching partner v
-          // must be matched in the meantime with an other vertex. Therefore,
-          // we try to match u with the representative v's cluster.
-          if ( _matching_state[u].load(std::memory_order_relaxed) == match_in_progress ) {
-            ASSERT( _matching_state[v] == matched );
-            const HypernodeID rep = cluster_ids[v];
-            success = joinCluster<has_fixed_vertices>(hypergraph,
-              u, rep, cluster_ids, contracted_nodes, fixed_vertices);
-          }
-        }
-        _rater.markAsMatched(u);
-        _rater.markAsMatched(v);
-        _matching_partner[u].store(u, std::memory_order_relaxed);
-        _matching_state[u].store(matched, std::memory_order_release);
-      }
-    }
-    return success;
-  }
-
-  template<bool has_fixed_vertices>
-  bool joinCluster(const Hypergraph& hypergraph,
-                   const HypernodeID u,
-                   const HypernodeID rep,
-                   vec<HypernodeID>& cluster_ids,
-                   HypernodeID& contracted_nodes,
-                   ds::FixedVertexSupport<Hypergraph>& fixed_vertices) {
-    ASSERT(rep == cluster_ids[rep]);
-    bool success = false;
-    const HypernodeWeight weight_of_u = hypergraph.nodeWeight(u);
-    const HypernodeWeight weight_of_rep = _cluster_weight[rep].load(std::memory_order_relaxed);
-    bool cluster_join_operation_allowed =
-      weight_of_u + weight_of_rep <= _context.coarsening.max_allowed_node_weight;
-    if constexpr ( has_fixed_vertices ) {
-      if ( cluster_join_operation_allowed ) {
-        cluster_join_operation_allowed = fixed_vertices.contract(rep, u);
-      }
-    }
-    if ( cluster_join_operation_allowed ) {
-      cluster_ids[u] = rep;
-      _cluster_weight[rep].fetch_add(weight_of_u, std::memory_order_relaxed);
-      ++contracted_nodes;
-      success = true;
-    }
-    return success;
-  }
-
   HypernodeID currentNumberOfNodesImpl() const override {
     return Base::currentNumNodes();
   }
@@ -512,12 +302,10 @@ class MultilevelCoarsener : public ICoarsener,
   using Base::_context;
   using Base::_timer;
   using Base::_uncoarseningData;
-  Rater _rater;
+  MultilevelVertexPairRater _rater;
   HypernodeID _initial_num_nodes;
   parallel::scalable_vector<HypernodeID> _current_vertices;
-  parallel::scalable_vector<AtomicMatchingState> _matching_state;
-  parallel::scalable_vector<AtomicWeight> _cluster_weight;
-  parallel::scalable_vector<AtomicID> _matching_partner;
+  ConcurrentClusteringData _clustering_data;
   int _pass_nr;
   utils::ProgressBar _progress_bar;
   bool _enable_randomization;
